refactor: fix misleading FileInfo objects returned by MarkdownParser
The MarkdownParser.ParseLinkedImages function returned FileInfo Objects but
it didn't consider the path of the Markdown source file. As a consequence
all image files in the returned list had the wrong path.

This change signature refactoring changes the return type to string, so
that only the image file names are returned.